yask: Move namespace to utils.py
